Change strategy for termination of resource limiting threads (#26)
Earlier, when some resource limiting thread(rlt) found the sandboxed
executable to be exceeding a certain resource usage limit, it was terminated
by the same rlt. Other rlts were also terminated by the same rltd.

However, now, when an rlt finds some limit to be exceeded, it terminates the
sandboxed executable as previously, but other rlts are not terminated by this rlt. All rlts
irrespective of the situation, are terminated by the main thread that
performs the wait.